Correct order in Product Docs navigation
## Type of change

### What should this PR do?

Set the order in the Product Docs navigation section so it makes sense

* Containers
* Libraries
* Registry
* Migration
* Administration
* chainctl Usage
* chainctl Reference

Also note that I am NOT updating timestamps since this is not a content
change.

### Why are we making this change?

Somehow adding the timestamps for Libraries bumped it to the very top
and it should not be above all the container related content.

### What are the acceptance criteria? 

review

### How should this PR be tested?

run locally and verify