Adjust to final URL for Chainguard Libraries for Java
## Type of change

### What should this PR do?

Update docs to the final URL.

### Why are we making this change?

New URL is more in line with naming of the product and should be used
going forward.

### What are the acceptance criteria? 

Review.

### How should this PR be tested?

URL must be live in production and a test should be run manually and
with a repo manager. Done by Manfred as soon as prod deployment is
ready.