Development containers (#2204)
[ ] Check if this is a typo or other quick fix and ignore the rest :)

## Type of change
<!-- Please be sure to add the appropriate label to your PR. -->
This PR updates the development vs production doc to use the current
language. We'd previously landed on "development" to refer to `-dev`
images and "production" to refer to non-dev images, but this language
was confusing as we use production to refer to private images.

I updated this language throughout Academy, and also updated 'images'
==> 'containers' where it made sense.

### What should this PR do?
<!-- Does this PR resolve an issue? Please include a reference to it.
-->
resolves https://github.com/chainguard-dev/internal/issues/4564

### Why are we making this change?
<!-- What larger problem does this PR address? -->
keeping our docs current

### What are the acceptance criteria? 
<!-- What should be happening for this PR to be accepted? Please list
criteria. -->
<!-- Do any stakeholders need to be tagged in this review? If so, please
add them. -->
Changes should make sense!

### How should this PR be tested?
<!-- What should your reviewer do to test this PR? Please list steps.
-->
A close read of the development containers (fka dev vs. production) doc
would be good, and then a less close read of the other changes would be
appreciated.

---------

Signed-off-by: Mark Drake <mark@chainguard.dev>
Signed-off-by: Mark Drake <33191761+SharpRake@users.noreply.github.com>