Adding simple example of how to subscribe to Chainguard Events (#2223)
[ ] Check if this is a typo or other quick fix and ignore the rest :)

## Type of change
<!-- Please be sure to add the appropriate label to your PR. -->
This PR translates a doc from the Registry Rockstar course to fit more
within our style for Chainguard Academy. It involves a brief example of
how to subscribe to an event and includes some details for validating
events.

### What should this PR do?
<!-- Does this PR resolve an issue? Please include a reference to it.
-->
This relates to [this
issue](https://github.com/chainguard-dev/internal/issues/4761) but
doesn't quite resolve it

### Why are we making this change?
<!-- What larger problem does this PR address? -->
Trying to make customers more aware of CloudEvents.

### What are the acceptance criteria? 
<!-- What should be happening for this PR to be accepted? Please list
criteria. -->
<!-- Do any stakeholders need to be tagged in this review? If so, please
add them. -->

### How should this PR be tested?
<!-- What should your reviewer do to test this PR? Please list steps.
-->
Should pass a simple tech test and be reviewed for content.

---------

Signed-off-by: Mark Drake <mark@chainguard.dev>