Updates to dfc user guide (#2228)
This PR brings some updates to the dfc guide. We're changing the title /
mentions to the tool from `Dockerfile Converter Tool` to `Dockerfile
Converter` for short, per request by Anushka Iyer.

There are also a few other small updates, added links to the project's
readme, and a section on using it as a Go library, linking back to the
readme with examples.

Preview Link:
https://deploy-preview-2228--ornate-narwhal-088216.netlify.app/chainguard/migration/dockerfile-conversion/