Adding new Custom Assembly docs (#2232)
[ ] Check if this is a typo or other quick fix and ignore the rest :)

## Type of change
<!-- Please be sure to add the appropriate label to your PR. -->
This PR adds two new docs that relate to Custom Assembly. One outlining
how to manage Custom Assembly with chainctl and another that shows off a
demo application that manages Custom Assembly with the API.

This also creates a new `Custom Assembly` bucket within the Containers
Features section to store these CA docs.

### What should this PR do?
<!-- Does this PR resolve an issue? Please include a reference to it.
-->
Resolves https://github.com/chainguard-dev/internal/issues/4884
Resolves https://github.com/chainguard-dev/internal/issues/4883

### Why are we making this change?
<!-- What larger problem does this PR address? -->
This is in anticipation of the upcoming Custom Assembly GA launch

### What are the acceptance criteria? 
<!-- What should be happening for this PR to be accepted? Please list
criteria. -->
<!-- Do any stakeholders need to be tagged in this review? If so, please
add them. -->
cc @sourabhkatti @tcnghia 

### How should this PR be tested?
<!-- What should your reviewer do to test this PR? Please list steps.
-->
Both docs should be tested and reviewed for content and clarity. Also,
please lmk if you think the structure I landed on makes sense. I know
the chainctl doc could potentially live under Chainctl Usage but I think
it's tied closely enough to the Custom Assembly feature that we don't
want it to be hard for users to find.

---------

Signed-off-by: Mark Drake <mark@chainguard.dev>