Combine and improve chainctl images pages (#2250)
Fixes https://github.com/chainguard-dev/internal/issues/4724 by
transforming the images page from more of a reference to something more
like a short tutorial.

This also deletes a page dedicated to the `chainctl images history`
sub-command and inserts the information into the main page. Does this
one need a redirect, @smythp and @ltagliaferri ? The history page was
released with a bit of fanfare as a cooperative effort and I would hate
to unnecessarily lose the goodwill. At the same time, it really doesn't
need it's own page.