Adjust Java repositories to naming convention
## Type of change

Follow up on the agreed standardizatio and the implementation in Python.

### What should this PR do?

Introduce consistent naming suitable for all ecosystems for the public,
chainguard, and grouped repositories similar to Python - now in Java.

### Why are we making this change?

Make it easier to understand and consistent. Also in prep for similar
things for JS again soon.

### What are the acceptance criteria? 

Review

### How should this PR be tested?

Not applicable.