Adding note to verifying sigs doc, tweaking Migs overview (#2309)
[ ] Check if this is a typo or other quick fix and ignore the rest :)

## Type of change
<!-- Please be sure to add the appropriate label to your PR. -->
This PR covers two small-ish changes. 

The first removes a section from the Migrations overview - I didn't
realize we already had a standalone migrations checklist.

The second adds a note to the verify signatures with Cosign doc, which
was recently updated. This update is being made in response to customer
feedback.

### What should this PR do?
<!-- Does this PR resolve an issue? Please include a reference to it.
-->
Resolves https://github.com/chainguard-dev/internal/issues/5072

### Why are we making this change?
<!-- What larger problem does this PR address? -->
Customer feedback, and my own sanity.

### What are the acceptance criteria? 
<!-- What should be happening for this PR to be accepted? Please list
criteria. -->
<!-- Do any stakeholders need to be tagged in this review? If so, please
add them. -->
The Signatures issue was raised to me by @kevinmdavis 

### How should this PR be tested?
<!-- What should your reviewer do to test this PR? Please list steps.
-->
Both of these are fairly small changes, so just a quick review would be
great.

---------

Signed-off-by: Mark Drake <mark@chainguard.dev>