Adding doc showing examples of how to auth using the SDK (#2307)
## Type of change
<!-- Please be sure to add the appropriate label to your PR. -->
This PR adds a doc outlining two examples included in the [SDK
repository](https://github.com/chainguard-dev/sdk/tree/main/examples/registry)
which demonstrate how to authenticate using the SDK.

### What should this PR do?
<!-- Does this PR resolve an issue? Please include a reference to it.
-->
Resolves https://github.com/chainguard-dev/internal/issues/4951

### Why are we making this change?
<!-- What larger problem does this PR address? -->
The SDK is currently not documented at all, this is part of an ongoing
effort to fix that.

### What are the acceptance criteria? 
<!-- What should be happening for this PR to be accepted? Please list
criteria. -->
<!-- Do any stakeholders need to be tagged in this review? If so, please
add them. -->
@wlynch put together the original examples. The doc should make sense
and work as written.

### How should this PR be tested?
<!-- What should your reviewer do to test this PR? Please list steps.
-->
a close read for content and a tech test would be appreciated!

[preview
link](https://deploy-preview-2307--ornate-narwhal-088216.netlify.app/chainguard/administration/sdk-authentication/)

---------

Signed-off-by: Mark Drake <mark@chainguard.dev>