Add Learning Labs section
### What should this PR do?

Add a new section in Education on the left hand nav for Learning Labs.

Fix https://github.com/chainguard-dev/internal/issues/5068

Also fix rendering of links to have a blue highlight color in topic pages (_index.md).

### Why are we making this change?

Each lab has a bunch of useful content and we need a place for our
viewers to access more information after the show and also create a spot
for finding the labs on the docs rather than on YouTube only. This will
also contribute to our traffic on the docs and be a good signal.

### What are the acceptance criteria? 

Passing review.

### How should this PR be tested?

Read and test links.