Add minimal info for chainver
### What should this PR do?

Add minimal docs for chainver .. what it is used for and how to get it. Usage
docs are part of the binary.

Fix https://github.com/chainguard-dev/internal/issues/5176

### Why are we making this change?

Another feature we supply to prospects that is ready to write about.

### What are the acceptance criteria? 

Confirm that the minimal info approach is ok.

### How should this PR be tested?

Not really necessary.