add iamguarded helm documentation (#2400)
[ ] Check if this is a typo or other quick fix and ignore the rest :)

## Type of change
This PR adds documentation for Chainguard's `iamguarded` Helm charts

### What should this PR do?
resolves https://github.com/chainguard-dev/internal/issues/5087

### Why are we making this change?
These are new to Chainguard and need documentation

### What are the acceptance criteria? 
Everything here has been checked in a Google doc by stakeholders, so all
we need to do is check for formatting errors and typos that may have
been introduced while moving the content over.

This builds beautifully locally using `npm run start`, but we should all
check that the indentation on the content under each ordered list item
in the preview.

### How should this PR be tested?
See above