Updating dfc docs (#2401)
This PR updates the DFC user guide so it is more in sync with what we'll
release in the DFC mini course.

The page is getting quite long, I'd like to hear your thoughts on
breaking it down into multiple docs.

Preview:
https://deploy-preview-2401--ornate-narwhal-088216.netlify.app/chainguard/migration/dockerfile-conversion/

Update July 14: added "Multi-Stage Builds" section with example and a
few other details after review.