chainguard-images: Add TLS networking requirements
Recently we had many support cases and requests around TLS
connectivity, as to what is supported what is required and how to test
and debug such connectivity.

Document the lowest common denominator of TLS connectivity as supported by:
- NIST FIPS
- CA/Browser Forum
- Let's Encrypt
- Fulcio
- All Chainguard FIPS and non-FIPS containers

It really boils down to a single hash, two signature algorithms, and
two pairs of TLS versions with a matching ciphersuite
string/group/curve.