Add info about pull token management (#2428)
## Type of change

Addition to docs. 

### What should this PR do?

Add info about pull token management

- In console and with chainctl
- Show list, show filtered list
- Delete expired tokens

Note that creating pull tokens for libraries in the console is not
supported and therefore not documented. Once this is done I will follow
up with a separate PR.

### Why are we making this change?

Feature exist but docs are missing.

Part of https://github.com/chainguard-dev/internal/issues/5111

### What are the acceptance criteria? 

Testing and PR review.

### How should this PR be tested?

With chainctl and console. Only deletion in console works so far ..
waiting for code merges and deployment for pull token creation in
console and new syntax and improved list and deletion support in
chainctl