POC - Feedback buttons - **DO NOT MERGE** (#2381)
[ ] Check if this is a typo or other quick fix and ignore the rest :)

## Type of change
<!-- Please be sure to add the appropriate label to your PR. -->
This PR adds a feedback button to each page. It is currently a WIP/POC
so please do not merge this yet.

The goal is to provide a direct avenue for readers to provide us with
feedback on specific Chainguard Academy resources.

### What should this PR do?
<!-- Does this PR resolve an issue? Please include a reference to it.
-->
resolves https://github.com/chainguard-dev/internal/issues/5246

### Why are we making this change?
<!-- What larger problem does this PR address? -->
We're making this change to make it easier for folks to provide us with
feedback and help us improve our docs faster.

### What are the acceptance criteria? 
<!-- What should be happening for this PR to be accepted? Please list
criteria. -->
<!-- Do any stakeholders need to be tagged in this review? If so, please
add them. -->
Buttons need to work as expected and effectively record feedback.

### How should this PR be tested?
<!-- What should your reviewer do to test this PR? Please list steps.
-->
The deploy preview likely won't be able to record any feedback.


[preview](https://deploy-preview-2381--ornate-narwhal-088216.netlify.app/chainguard/chainguard-images/about/what-chainguard-will-build/)

---------

Signed-off-by: Mark Drake <mark@chainguard.dev>
Signed-off-by: Mark Drake <33191761+SharpRake@users.noreply.github.com>