Add Kubernetes assumable identities example (#2476)
[ ] Check if this is a typo or other quick fix and ignore the rest :)

## Type of change
<!-- Please be sure to add the appropriate label to your PR. -->
This PR adds a doc on creating an assumable identity for a Kubernetes
pod.

### What should this PR do?
<!-- Does this PR resolve an issue? Please include a reference to it.
-->

Adds a new article that describes how to create an assumable identity
for a Kubernetes pod.

### Why are we making this change?
<!-- What larger problem does this PR address? -->

I have been asked about this by a number of customers in the past. It
also comes up a lot internally and we don't have this consolidated
anywhere.

### What are the acceptance criteria? 
<!-- What should be happening for this PR to be accepted? Please list
criteria. -->
<!-- Do any stakeholders need to be tagged in this review? If so, please
add them. -->

🤷 

### How should this PR be tested?
<!-- What should your reviewer do to test this PR? Please list steps.
-->

Create a Kubernetes cluster and then follow the steps. You can do this
locally with a tool like `kind`.

---------

Signed-off-by: Rob Best <robertbest89@gmail.com>
Co-authored-by: Matthew Helmke <matthew.helmke@chainguard.dev>