Documentation for the Feedback Widget (#2507)
[ ] Check if this is a typo or other quick fix and ignore the rest :)

## Type of change
<!-- Please be sure to add the appropriate label to your PR. -->
New platform doc outlining the functionality of the feedback widget.

### What should this PR do?
<!-- Does this PR resolve an issue? Please include a reference to it.
-->

### Why are we making this change?
<!-- What larger problem does this PR address? -->
I added the feedback widget last week without including detailed info on
how it works. This is meant to provide a general overview without giving
away too many details.

### What are the acceptance criteria? 
<!-- What should be happening for this PR to be accepted? Please list
criteria. -->
<!-- Do any stakeholders need to be tagged in this review? If so, please
add them. -->
Doc should serve as a starting point for folks looking to understand the
feedback widget.

### How should this PR be tested?
<!-- What should your reviewer do to test this PR? Please list steps.
-->
Just a read, this isn't going on Academy so it doesn't need to adhere to
the style guide.

Signed-off-by: Mark Drake <mark@chainguard.dev>