Add docs for library pull token creation in console (#2596)
## Type of change

Adding docs for libraries

### What should this PR do?

Add info on how to create and modify a pull token in the console so it
can be used with libraries.

### Why are we making this change?

Wrongly thought that this is not possible and never heard from anyone
yet to the contrary.

Fix https://github.com/chainguard-dev/internal/issues/5365

### What are the acceptance criteria? 

Precommit hooks passed.

Pass review.

### How should this PR be tested?

Follow documented steps