Adding Ko guide in new Tooling section (#2591)
## Type of change
This PR adds a new section in Chainguard Academy dedicated to container
building tools, or "last mile tools" as we initially called them. To
avoid confusion with future tools we may (or may not) build, we'll keep
it simple (also respecting menu sizing) and just call it "Tooling" for
the menu. It also adds the first tutorial in this section, using Ko to
build Go containers with Chainguard base images.

### What should this PR do?
This work is described
[here](hhttps://github.com/chainguard-dev/internal/issues/5198).

### Why are we making this change?
This is part of an initiative called "last mile container tools" to
bring resources on how to use tools other than Docker to build container
images.

### What are the acceptance criteria? 
- Tech test passes
- Content follows our technical writing guidelines
- Menu and titles look good and make sense under our current structure

### How should this PR be tested?
- Live preview:
https://deploy-preview-2591--ornate-narwhal-088216.netlify.app/chainguard/chainguard-images/tooling/building-go-containers-with-ko/

Disclaimer: I was not able to build the site locally, currently getting
several errors from the Hugo build / serve, so I am depending on the
live preview to see how it will look when published.