Document 'Disable URL Normalization' option for Artifactory APK repositories (#2624)
[ ] Check if this is a typo or other quick fix and ignore the rest :)

## Type of change
<!-- Please be sure to add the appropriate label to your PR. -->

Documentation.

### What should this PR do?
<!-- Does this PR resolve an issue? Please include a reference to it.
-->

Document an additional required setting when configuring a remote
repository for our APK repositories in Artifactory.

### Why are we making this change?
<!-- What larger problem does this PR address? -->

Without this setting, Artifactory fails to pull packages with a `+`
character in the checksum.

We use the checksum as the object name in Cloudflare and Artifactory's
normalization seemingly mangles the URL encoding when it does its
normalization. This results in a 403.

### What are the acceptance criteria? 
<!-- What should be happening for this PR to be accepted? Please list
criteria. -->
<!-- Do any stakeholders need to be tagged in this review? If so, please
add them. -->

🤷 

### How should this PR be tested?
<!-- What should your reviewer do to test this PR? Please list steps.
-->

1. Follow the steps as they are currently documented to set up a remote
repository for a private or public APK repository.
2. Then try to add a package with a `+` in the checksum.

Here's an example. This must be ran from a `linux/amd64` container:
```
apk add curl=8.15.0-r5
```

You can also try to fetch it directly with curl:
```
curl -u "${ARTIFACTORY_CREDENTIALS}" 'https://chainguard.jfrog.io/artifactory/rob-best-virtualapk-chainguard/x86_64/curl-8.15.0-r5.apk'
```

You should get an error like:
```
{
  "errors" : [ {
    "status" : 404,
    "message" : "Forbidden; Path: 'rob-best-virtualapk-chainguard:x86_64/curl-8.15.0-r5.apk'"
  } ]
}
```

Enable 'Disable URL Normalization' and observe that the issue is
resolved.