Add auth with MS Entra ID to docs (#2499)
EDIT: As noted in [this
comment](https://github.com/chainguard-dev/edu/pull/2499#issuecomment-3200802277),
I signed up for a free one month trial account for Microsoft Entra ID
and tested this PR myself. I got it to work, but am not confident as a
Entra ID user, so that's not a glowing recommendation. This was on top
of collaborative testing attempts with Philippe, as noted in that
comment.

Fixes https://github.com/chainguard-dev/internal/issues/5234