Google csp (#2634)
[ ] Check if this is a typo or other quick fix and ignore the rest :)

## Type of change
Updated the CSP settings. Currently the edu site is failing to load many
references to connect-src, script-src, img-src and frame-src. I added
all the scripts that were failing on the edu site.

### What should this PR do?
This PR should resolve current broken CSP console errors.

### Why are we making this change?
To allow tag manger to track user visits to edu site and allow scripts
on the site to work correctly.

### What are the acceptance criteria? 
The CSP console errors should be cleared up when loading the edu site

### How should this PR be tested?
You can pull this branch down locally test if there are any CSP errors
in the console. If there is a staging site this should also work there
as well.