Update custom-assembly.md with additional perms and --save-as updates (#2719)
[ ] Check if this is a typo or other quick fix and ignore the rest :)

## Type of change
<!-- Please be sure to add the appropriate label to your PR. -->

### What should this PR do?
Updates CA permissions information to align with current status of the
feature.

### Why are we making this change?
1. build_history.list was not included in the example role, so an image
could be customized in place, but users with this example role wouldn't
be able to actually view the build history to explicitly confirm if
their customization was successful.

2. with --save-as, repo.create is required because a net new image repo
is created. I added some additional context about it towards the top of
the permissions section and in the example role.

### What are the acceptance criteria? 
1. Review added language describing --save-as for readability.
2. Confirming custom role requiring repo.create for --save-as.
3. Confirm user can view build_history with the suggested custom role
update.
@SharpRake

### How should this PR be tested?
1. Run the existing custom role example and attempt to view
build_history and do --save-as. The expected result is that both
operations will fail.
2. Run the suggested custom role in this PR and attempt to view
build_history and do --save-as. The expected result is that it both
operations will work.

---------

Signed-off-by: Madhav Mehta <madhav.mehta@chainguard.dev>
Signed-off-by: Mark Drake <33191761+SharpRake@users.noreply.github.com>
Co-authored-by: Mark Drake <33191761+SharpRake@users.noreply.github.com>