Updates to AI security doc. (#2740)
This PR improves the verification of the download by adding the name of
action (otherwise any action could have signed, including non-chainguard
ones).

I deleted section with verification script, as there is no indication of
where to get this file and we've already verified it's authenticity. If
the script exists and works, we should add it back.

I also deleted the container section as verification failed and I'm
unsure of the use case -- why would you pull the container rather than
the .gz? Happy to add back and test if we can get this working and
there's a valid use case.

## Type of change
Update to docs following testing.

### What should this PR do?
Improve verification and remove unused flow.

### Why are we making this change?
Allow users to effectively verify download.

### What are the acceptance criteria? 
@ltagliaferri to check this is correct!

### How should this PR be tested?
Please test verification passes.

Signed-off-by: Adrian Mouat <adrian@chainguard.dev>