Update package-model.md (#2789)
Add information about the package retention policy for wolfi & private
apk repos to the package repository docs.

[ ] Check if this is a typo or other quick fix and ignore the rest :)

## Type of change
documentation

### What should this PR do?
Adds information on our package retention policy to our docs.  

### Why are we making this change?
Currently, this information exists as an announcement in the Wolfi
GitHub (https://github.com/orgs/wolfi-dev/discussions/63023), but a few
customers want to reference it in their deals, so we also need it in our
docs.

### What are the acceptance criteria? 
If the content is accurate & consistent with Mark's vision for these
docs.

### How should this PR be tested?
N/A - this describes a process, and is not testable by customers.

---------

Signed-off-by: John Slack <johncslack@gmail.com>
Signed-off-by: Mark Drake <33191761+SharpRake@users.noreply.github.com>
Co-authored-by: Mark Drake <33191761+SharpRake@users.noreply.github.com>