Add a generic assumable identity page for AWS.
I have received feedback from customers and colleagues that our AWS
Lambda example is hard to follow because it uses Terraform and doesn't
really describe fundamentally how our AWS assuamble identities work.

Our EC2 example is better but it is difficult to translate to other AWS
services because it relies on the EC2 instance metadata service to fetch
credentials.

This change adds a generic page that should be relevant and translate
well for most use cases.

Firstly, it provides generic examples of setting up an identity that
will work for essentially any service, as long as they are using an IAM
user or assuming an IAM role.

Secondly, it actually tries to outline how our approach for AWS
assumable identities works and why it differs from other platforms.
Which I think is important to know while we are requiring users to run
arcane curl commands.

Thirdly, it uses `chainctl` examples which are much easier to grok, IMO.

I've also modified the ordering of the AWS pages so they should sort
together in the nav bar.