github-identity: Add chainctl example (#2807)
I've received a lot of feedback from customers and colleagues that this
page is hard to follow because it exclusively uses Terraform for its
examples.

This change adds an example of using `chainctl` to create the identity.

I've also modified the Terraform example to be less procedural. I think
stepping through all the files and operations just obfuscates the
information that most people care about: what resources do I need to
configure?

I think its safe to assume that if someone is looking for a Terraform
example, then they are comfortable with defining a provider, setting up
a Terraform code base and applying the configuration. Using a snippet
communicates the relevant information in a short and digestable way.