assumable-ids: Add generic examples of assuming an identity
We provide generic examples of creating an identity, but no tangible
examples of assuming one.

Most importantly, this adds an example of doing the exchange entirely
via the API. This comes up a lot with customers who are writing their
own implementations with our APIs and don't want to shell out to `chainctl`
for authentication.