assumable-ids: Add generic examples of assuming an identity (#2811)
We provide generic examples of creating an identity, but no tangible
examples of assuming one.

This change adds generic instructions on how to assume an identity.

Most importantly, it adds an example of doing the exchange entirely via
the API. This comes up a lot with customers who are writing their own
implementations with our APIs and don't want to shell out to `chainctl`
for authentication.

---------

Signed-off-by: Mark Drake <33191761+SharpRake@users.noreply.github.com>
Co-authored-by: Mark Drake <33191761+SharpRake@users.noreply.github.com>